+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.5
+
+* Written tests for Router v2 (WIP)
+* You can use PtonV1 and PtonV2 both for Router v1 and v2
+
 ## 0.0.4
 
 * Written some tests to demonstrate working state